10477 Move tax on other income reads into companion obj
